review: move yield and zc token tests to testing package
